requirements.txt
numpy
pandas
scipy
scikit-learn
xgboost
pyyaml
requests
tqdm
matplotlib
seaborn
torch
transformers
peft
trl
datasets
accelerate
geopandas
shapely
vllm
